+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Add `externalresources` resource that binds `read-default-catalogs` and `read-releases` roles for any subject with org-namespace access.
 - Add creation of `read-default-catalogs` Role.
 - Add creation of `read-releases` ClusterRole.
 - In the `orgpermissions` controller, remove logging for non-operations and add namespace info to log messages.